Move accordion to separate file.
