+++ jupyter-js-widgets/src/phosphor/phosphor_widgets.ts
@@ -41,10 +41,6 @@ import {
     Message
 } from 'phosphor/lib/core/messaging';
 
-import {
-  VectorWithSelection
-} from './phosphorVectorSelection';
-
 /**
  * The class name added to TabPanel instances.
  */
@@ -550,276 +546,3 @@ namespace Private {
 }
 
 
-/**
- * A panel that supports a collapsible header, made from the widget's title.
- * Clicking on the title expands or contracts the widget.
- */
-export
-class Collapse extends Widget {
-  constructor(options: Collapse.IOptions) {
-    super(options)
-    this._header = new Widget();
-    this._header.node.addEventListener('click', this)
-
-    let layout = new PanelLayout();
-    layout.addWidget(this._header);
-
-    if (options.widget) {
-      this.widget = options.widget;
-    }
-
-    this.layout = layout;
-  }
-
-  dispose() {
-    if (this.isDisposed) {
-      return;
-    }
-    super.dispose();
-    this._header = null;
-    this._widget = null;
-  }
-
-  get widget() {
-    return this._widget;
-  }
-
-  set widget(widget: Widget) {
-    let layout = this.layout as PanelLayout;
-    if (this._widget) {
-      this._widget.title.changed.disconnect(this._onTitleChanged, this);
-      layout.removeWidget(this._widget);
-    }
-    if (this.collapsed) {
-      widget.hide();
-    } else {
-      widget.show();
-    }
-    widget.title.changed.connect(this._onTitleChanged, this);
-    this._onTitleChanged(widget.title);
-    layout.addWidget(widget);
-    this._widget = widget;
-  }
-
-  get collapsed() {
-    return this._collapsed;
-  }
-
-  set collapsed(value: boolean) {
-    // TODO: should we have this check here?
-    if (value === this._collapsed) {
-      return;
-    }
-    if (value) {
-      this._collapse();
-    } else {
-      this._uncollapse();
-    }
-  }
-
-  toggle() {
-    this.collapsed = !this.collapsed;
-  }
-
-  collapseChanged: ISignal<Collapse, void>;
-
-  private _collapse() {
-    this._collapsed = true;
-    this._widget.hide();
-    this.collapseChanged.emit(void 0);
-  }
-  private _uncollapse() {
-    this._collapsed = false;
-    this._widget.show();
-    this.collapseChanged.emit(void 0);
-  }
-
-  /**
-   * Handle the DOM events for the Collapse widget.
-   *
-   * @param event - The DOM event sent to the panel.
-   *
-   * #### Notes
-   * This method implements the DOM `EventListener` interface and is
-   * called in response to events on the panel's DOM node. It should
-   * not be called directly by user code.
-   */
-  handleEvent(event: Event): void {
-    switch (event.type) {
-    case 'click':
-      this._evtClick(event as MouseEvent);
-      break;
-    }
-  }
-
-  private _evtClick(event: MouseEvent) {
-    this.toggle();
-  }
-
-  /**
-   * Handle the `changed` signal of a title object.
-   */
-  private _onTitleChanged(sender: Title): void {
-    this._header.node.textContent = this._widget.title.label;
-  }
-
-  _collapsed: boolean;
-  _widget: Widget;
-  _header: Widget;
-}
-
-export
-namespace Collapse {
-  export
-  interface IOptions extends Widget.IOptions {
-    widget: Widget;
-  }
-}
-
-// Define the signals for the `Widget` class.
-defineSignal(Collapse.prototype, 'collapseChanged');
-
-export
-class AccordionLayout extends PanelLayout {
-  private _widgets = VectorWithSelection
-}
-
-
-/**
- * A panel that supports a collapsible header, made from the widget's title.
- * Clicking on the title expands or contracts the widget.
- */
-export
-class Accordion extends Panel {
-
-
-  // Logic that if a widget is collapsed/not collapsed, it updates the
-  // currentWidget and currentIndex.
-
-  /**
-   * Add a widget to the end of the accordion.
-   *
-   * @param widget - The widget to add to the accordion.
-   *
-   * #### Notes
-   * The widget will be wrapped in a CollapsedWidget.
-   */
-  addWidget(widget: Widget): void {
-    super.addWidget(this._wrapWidget(widget));
-  }
-
-  /**
-   * Insert a widget at the specified index.
-   *
-   * @param index - The index at which to insert the widget.
-   *
-   * @param widget - The widget to insert into to the accordion.
-   *
-   * #### Notes
-   * If the widget is already contained in the panel, it will be moved.
-   */
-  insertWidget(index: number, widget: Widget): void {
-    super.insertWidget(index, this._wrapWidget(widget));
-  }
-
-  _wrapWidget(widget: Widget) {
-    let collapse = new Collapse({ widget });
-    widget.disposed.connect(() => collapse.dispose())
-    return collapse;
-  }
-
-  /**
-   * A signal emitted when the current tab is changed.
-   *
-   * #### Notes
-   * This signal is emitted when the currently selected tab is changed
-   * either through user or programmatic interaction.
-   *
-   * Notably, this signal is not emitted when the index of the current
-   * tab changes due to tabs being inserted, removed, or moved. It is
-   * only emitted when the actual current tab node is changed.
-   */
-  currentChanged: ISignal<TabPanel, TabPanel.ICurrentChangedArgs>;
-
-  /**
-   * Get the index of the currently selected tab.
-   *
-   * #### Notes
-   * This will be `-1` if no tab is selected.
-   */
-  get currentIndex(): number {
-    return this._currentIndex;
-  }
-
-  /**
-   * Set the index of the currently selected tab.
-   *
-   * #### Notes
-   * If the index is out of range, it will be set to `-1`.
-   */
-  set currentIndex(value: number) {
-    this._currentIndex = value;
-  }
-
-  /**
-   * Get the currently selected widget.
-   *
-   * #### Notes
-   * This will be `null` if there is no selected tab.
-   */
-  get currentWidget(): Widget {
-    let title = this._tabBar.currentTitle;
-    return title ? title.owner as Widget : null;
-  }
-
-  /**
-   * Set the currently selected widget.
-   *
-   * #### Notes
-   * If the widget is not in the panel, it will be set to `null`.
-   */
-  set currentWidget(value: Widget) {
-    this._tabBar.currentTitle = value ? value.title : null;
-  }
-
-  _currentIndex: number;
-  _currentWidget: Widget;
-}
-
-
-// Define the signals for the `Widget` class.
-defineSignal(Accordion.prototype, 'currentChanged');
-
-export
-namespace Accordion {
-  export
-  type IOptions = Panel.IOptions;
-
-  /**
-   * The arguments object for the `currentChanged` signal.
-   */
-  export
-  interface ICurrentChangedArgs {
-    /**
-     * The previously selected index.
-     */
-    previousIndex: number;
-
-    /**
-     * The previously selected widget.
-     */
-    previousWidget: Widget;
-
-    /**
-     * The currently selected index.
-     */
-    currentIndex: number;
-
-    /**
-     * The currently selected widget.
-     */
-    currentWidget: Widget;
-  }
-
-
-}
